Fixing the Nomad config file
We now have two 'environments' - the default config in the root of the
map, and the production config. We switch between these with the
NOMAD_ENV environment variable - setting this to "prod" will activate
the production config.

Normally, Nomad (and, by extension, Frodo) will use the default config
provided (i.e. port 3000, 'cljs-repl?' true etc), but if we switch it to
prod, it'll look in the PORT env-var for the port, and turn the
cljs-repl off.